backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
celery[redis]==5.4.0
redis==5.2.0
google-generativeai==0.8.3
aiogram==3.15.0
python-multipart==0.0.22
httpx==0.28.1
